+++ NEWS.md
@@ -10,6 +10,14 @@
   string instead of only a hash (e.g. `route(...).to('posts#index')` instead of
   `route(...).to(controller: 'posts', action: 'index')`).
 
+* The `ensure_inclusion_of` matcher now works with a decimal column.
+
+* Fix association matchers on Rails 3 so they work when used in conjunction with
+  a submatcher such as #order and an association that has :include on it.
+
+* Fix a bug where `validate_uniqueness_of` would fail if the attribute under
+  test had a limit of fewer than 16 characters.
+
 # v 2.4.0
 
 * Fix a bug with the `validate_numericality_of` matcher that would not allow the